+++ ConstantIndexOrVertexBuffer.h
@@ -8,7 +8,7 @@ protected:
 	ComPtr<ID3D12Resource2> tempBuffer;
 public:
 	using SinglePtr::SinglePtr;
-	ConstantIndexOrVertexBuffer(ID3D12Device10* device, UINT size, UINT vertexSize);
+	ConstantIndexOrVertexBuffer(ID3D12Device10* device, uint64_t size, uint8_t vertexSize);
 	void Copy(ID3D12Device10* device, ID3D12GraphicsCommandList7* list, BYTE* pData, D3D12_RESOURCE_STATES state);
 	void ReleaseTemp();
 };
