+++ RootSignature.cpp
@@ -1,6 +1,6 @@
 #include "RootSignature.h"
 
-RootSignature::RootSignature(ID3D12Device10* device, UINT numParameters, const D3D12_ROOT_PARAMETER* pParams, UINT numStaticSamplers, const D3D12_STATIC_SAMPLER_DESC* pSamplers, D3D12_ROOT_SIGNATURE_FLAGS flags, D3D_ROOT_SIGNATURE_VERSION version)
+RootSignature::RootSignature(ID3D12Device10* device, uint16_t numParameters, const D3D12_ROOT_PARAMETER* pParams, uint16_t numStaticSamplers, const D3D12_STATIC_SAMPLER_DESC* pSamplers, D3D12_ROOT_SIGNATURE_FLAGS flags, D3D_ROOT_SIGNATURE_VERSION version)
 {
 	descriptor.Init(numParameters, pParams, numStaticSamplers, pSamplers, flags);
 	ComPtr<ID3DBlob> signature;
